explicit static_cast added
git-svn-id: svn://svn.r-forge.r-project.org/svnroot/rcpp@120 edb9625f-4e0d-4859-8d74-9fd3b1da38cb

+++ pkg/src/RcppExample.cpp
@@ -437,7 +437,7 @@ RcppExport SEXP RcppVectorExample(SEXP vector) {
 	std::vector<double> res(n);
 	
 	for (int i=0; i<n; i++) {
-	    res[i] = sqrt(vec(i));
+	    res[i] = sqrt(static_cast<double>(vec(i)));
 	}
 
 	// Build result set to be returned as a list to R.
